@@ -11,10 +11,13 @@
   "dependencies": {
     "express": "^4.16.2",
     "express-session": "^1.15.6",
+    "moment": "^2.22.0",
     "npm": "^5.8.0",
-    "react": "^16.2.0",
+    "react": "^16.3.0",
     "react-addons-test-utils": "^0.14.7",
-    "react-dom": "^16.2.0"
+    "react-dom": "^16.2.0",
+    "redux": "^3.7.2",
+    "uuid": "^3.2.1"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
