@@ -1,9 +1,23 @@
 var React = require('react');
+import { connect } from 'react-redux';
+import {bindActionCreators} from 'redux';
 
 import ConversationList  from 'ConversationList';
 import AskQuestion from 'AskQuestion';
 
-class ChatBox extends React.Component{
+import addToConversation from "actions";
+
+const mapStateToProps = function (state) {
+  return {
+    conversationList: state.conversationList
+  };
+}
+
+function mapDispatchToProps(dispatch) {
+    return bindActionCreators({addToConversation:addToConversation}, dispatch);
+}
+
+export class ChatBox extends React.Component{
   render() {
     var {dispatch} = this.props;
     return (
@@ -16,4 +30,4 @@ class ChatBox extends React.Component{
   }
 }
 
-module.exports = ChatBox;
+export default connect(mapStateToProps, mapDispatchToProps)(ChatBox);
